I'm going to go with a class structure for the interface pages.
I should consider using the @staticmethod decorator so the class does
not have to keep referring to its own method's with self.X. 

I am now going to experiment and see if the forget method can be called
and then the pack again. If so all 5 pages can be instantiated at once
and you just need to hop between them.